// src/generated/prisma/enums.ts
export const SocialPlatform = {
  FACEBOOK: 'FACEBOOK',
  LINKEDIN: 'LINKEDIN'
} as const

export type SocialPlatform = (typeof SocialPlatform)[keyof typeof SocialPlatform]
// ...
export const PostStatus = {
  DRAFT: 'DRAFT',
  MANUAL_PENDING: 'MANUAL_PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  SCHEDULED: 'SCHEDULED',
  PENDING_APPROVAL: 'PENDING_APPROVAL',
  REJECTED: 'REJECTED',
  APPROVED: 'APPROVED',
  PROCESSING: 'PROCESSING',
  PUBLISHED: 'PUBLISHED',
  PARTIALLY_PUBLISHED: 'PARTIALLY_PUBLISHED',
  FAILED: 'FAILED',
  BLOCKED: 'BLOCKED'
} as const

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus]


export const PostTargetStatus = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED',
  OVERDUE: 'OVERDUE',
  QUEUED: 'QUEUED',
  PROCESSING: 'PROCESSING',
  PUBLISHED: 'PUBLISHED',
  BLOCKED: 'BLOCKED'
} as const

export type PostTargetStatus = (typeof PostTargetStatus)[keyof typeof PostTargetStatus]